+++ CHANGELOG.md
@@ -52,6 +52,7 @@ IMPROVEMENTS:
 * builder/virtualbox: Nice errors if Packer can't write to
   the output directory.
 * builder/virtualbox: ISO is ejected prior to export.
+* builder/vmware: Can now specify path to the Fusion application. [GH-677]
 * provisioner/puppet-masterless: Can now specify a `manifest_dir` to
   upload manifests to the remote machine for imports. [GH-655]